Add simple notification workflow
This adds an api on the `HubController` which UI clients can use to send
notifications. These notifications, if there are interested subscribers,
should then trigger a notification to the callback configed in the
subscription.

Created a base model class that can be used to get common functionality
like printing debug info about the model.